requirements.txt
kopf
kubernetes
pyyaml


